Display the help text in the status window when no other line is shown.
Debugging is enabled only when DEBUG is defined and with the new
-D command-line switch.

Added command line options -V to validate a file and -S to validate
and write PGN formatted output to stdout. If there is a parse error
then cboard will return 1, otherwise 0.

Added command line option -N to disable initializing the chess engine.
This allows for two human players.