Fix the default engine command (gnuchess --xboard). Requires GNU Chess 5.07.
Added common.h to the archive.

Fix playing with engine.

Added .data to GAME. This is a void * that can hold any data for the
associated game. cboard uses it for engine (and later ICS) file
descriptors along with the pid and engine status.

More than one game can use a chess engine. If another game is started
which uses a chess engine a new engine process will be started.
Switching between games works as normal.