A more complete clock. Time controls are now supported so you can do stuff like: 20/4m30s 10/2m sd/3m +5, to have the first time control be 20 moves in 4 minutes 30 seconds then the second time control then sudden death in 3 minutes. The +5 is a move increment. This adds 5 seconds after each move.
All time values may have 'h', 'm' or 's' appended to indicate hours,
minutes or seconds. If not specified, seconds will be used.

Each player can have it's own settings by prefixing the clock
specification with either 'w' for white or 'b' for black. If neither
are specified both clocks will be modified.

A blitz game: 5m+12s to set a 5 minute game with a 12 second move
increment.

During a game: w+1m to give white an additional 1 minute.

During a game: w++4 or b++4 to set white or black move increment to 4
seconds.

During a game: ++4 to set both white an black move increment. Can also
be set to 0 to disable.

The status window has also been updated to show the number of moves
until the next time control, the next time control and the move
increment.

This is only working for human/human play. Working on chess engine
support...