Add "engine_timeout" configuration parameter.
This is the amount of seconds to wait for the engine to spawn. Rather
than waiting for a hardcoded 1 second then sending a test signal to the
engine process, wait for a user-defined amount of time (default 10).
Fixes the "no such process" error on a machine with a higher load
average when starting a new game.